Added complete functionality for global recomienda and highlight

+++ app/controllers/sections_controller.rb
@@ -8,28 +8,49 @@ class SectionsController < ApplicationController
 	end
 	def set_highlight_and_recomendations
 		article = Article.find(params[:highlight])
-		article.update_attributes(highlight: true)
-
+		
 		articles =  Article.where(id: params[:all_recommendations])
 		articles_in_highlights = SectionHighlight.where(article_id: params[:all_recommendations], section: article.articable)
-		articles.each do |this_article|
-			created = false
-			articles_in_highlights.each do |a_in_h|
-				if this_article == a_in_h.article
-					created = true
+		if articles.count >= 4
+			articles.each do |this_article|
+				created = false
+				articles_in_highlights.each do |a_in_h|
+					if this_article == a_in_h.article
+						created = true
+					end
+				end
+				if created == false
+					SectionHighlight.create(article_id: this_article.id, section_id: this_article.articable.id )
 				end
 			end
-			if created == false
-				SectionHighlight.create(article_id: this_article.id, section_id: this_article.articable.id )
+
+			if articles_in_highlights.count >= 4
+				get_first_h = SectionHighlight.where(section: article.articable).first(articles_in_highlights.count - 3)
+				get_first_h.each do |h|
+					h.destroy
+				end 
 			end
 		end
 
-		if articles_in_highlights.count >= 4
-			get_first_h = SectionHighlight.where(section: article.articable).first(articles_in_highlights.count - 3)
-			get_first_h.each do |h|
-				h.destroy
-			end 
+		current_articles = Article.where(articable: article.articable)
+		current_articles.each do |a|
+			p "UPDATING"
+			p a.name
+			a.update_attributes(highlight: false)
+			p a.highlight
 		end
+		article.update_attributes(highlight: true)
+		p "ARTICLE highlight"
+		p article.name
+		p article.highlight
+
+		current_articles.each do |a|
+			p "UPDATING"
+			p a.name
+			a.update_attributes(highlight: false)
+			p a.highlight
+		end
+
 			#SectionHighlight.create(article_id: this_article.id, section_id: this_article.articable.id )
